Simplifying the generation script
