+++ .cargo/config
@@ -1,2 +1,2 @@
 [alias]
-lv2-sys-bindgen = "run -p lv2-sys-bindgen --"
+make-bindings = "run -p lv2-sys-bindgen --"
