+++ sys/bindgen/Cargo.toml
@@ -1,11 +1,14 @@
 [package]
 name = "lv2-sys-bindgen"
 version = "0.1.0"
-authors = ["Yruama_Lairba <yruama_lairba@hotmail.com>"]
+authors = ["Yruama_Lairba <yruama_lairba@hotmail.com>", "Jan-Oliver 'Janonard' Opdenhövel <jan.opdenhoevel@protonmail.com>"]
 edition = "2018"
+license = "MIT OR Apache-2.0"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+description = "Generate Rust bindings of the LV2 C API"
+readme = "README.md"
+repository = "https://github.com/RustAudio/rust-lv2"
 
 [dependencies]
 bindgen = "0.53.1"
-
+clap = "2.33.0"
