--- sys/bindgen/enum_test.h
@@ -1,1 +1,0 @@
-enum test {TEST};
